refactor(storage) : rename storage folder
- represent folder to be related to state
- make clear local folder to be related to browser
