+++ src/components/quiz/DetailQuiz.tsx
@@ -9,7 +9,7 @@ import { QuizStatus } from '../../server-action/backend/quiz/type';
 import { isHttpException, isServerActionError } from '../../server-action/backend/common/util';
 import { DetailQuizDTO, QuizContextDTO } from '../../server-action/backend/quiz/dto';
 import {  getPaintingAction } from '../../server-action/backend/painting/api';
-import { getQuizStatus, saveQuizStatus } from '../../storage/local/quiz';
+import { getQuizStatus, saveQuizStatus } from '../../state/browser/quiz';
 import { ErrorModal } from '../modal/ErrorModal';
 import { QuizMenu } from './QuizMenu';
 interface DetailQuizProps {
